pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "guardrail-calibration-eval"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3",
    "transformers>=4.44",
    "bitsandbytes>=0.43",
    "accelerate>=0.30",
    "scipy>=1.13",
    "numpy>=1.26",
    "pandas>=2.2",
    "pyarrow>=16",
    "pyyaml>=6.0",
    "openai>=1.30",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "scikit-learn>=1.5",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100",
    "pytest>=8.2",
    "pytest-xdist>=3.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
